Refacator: 멤버 도메인으로 언어 설정 위치 수정
