--- src/main/java/org/lxdproject/lxd/member/repository/LanguageSettingRepository.java
@@ -1,4 +1,0 @@
-package org.lxdproject.lxd.member.repository;
-
-public class LanguageSettingRepository {
-}
